Add link to Vite.js SSR example with Svelte from vite-plugin-svelte
The [@sveltejs/vite-plugin-svelte](https://github.com/sveltejs/vite-plugin-svelte) which is used in SvelteKit contains a Vite.js SSR example with Svelte. Since there are links to Vue and React projects, it would help folks who want to understand how Vite.js and SSR work with Svelte to have a link to a Svelte example that’s part of a real-world project. (It took me a while to track this down so this would have saved me some time had I seen a link to it here.)